% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gcv.R
\name{compute_gcv}
\alias{compute_gcv}
\title{Function to compute generalized cross-validation score}
\usage{
compute_gcv(
  Y,
  delta,
  C,
  mu_hat,
  b_hat,
  lambda,
  Pen,
  family = c("lognormal", "loglogistic")
)
}
\arguments{
\item{Y}{Survival time}

\item{delta}{Censoring indicator (1 = event)}

\item{C}{Design matrix}

\item{mu_hat}{Estimated linear predictor}

\item{b_hat}{Estimated scale parameter}

\item{lambda}{A chosen nonnegative smoothing parameter}

\item{Pen}{A chosen penalty matrix}

\item{family}{"lognormal" or "loglogistic"}
}
\value{
Generalized cross-validation score
}
\description{
Function to compute generalized cross-validation score
}
