% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_loglik.R
\name{penalized_loglik}
\alias{penalized_loglik}
\title{Function to compute the negative log-likelihood}
\usage{
penalized_loglik(
  params,
  Y,
  delta,
  X,
  family = c("lognormal", "loglogistic"),
  lambda,
  Pen
)
}
\arguments{
\item{params}{A vector of parameters (gamma + b + sigma)}

\item{Y}{Survival time}

\item{delta}{Censoring indicator (1 = event)}

\item{X}{Design matrix}

\item{family}{"lognormal" or "loglogistic"}

\item{lambda}{A chosen nonnegative smoothing parameter}

\item{Pen}{A chosen penalty matrix}
}
\value{
Negative penalized log-likelihood
}
\description{
Function to compute the negative log-likelihood
}
