% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_score.R
\name{penalized_score}
\alias{penalized_score}
\title{Function to compute the gradient (score equations)}
\usage{
penalized_score(
  params,
  Y,
  delta,
  X,
  family = c("lognormal", "loglogistic"),
  lambda,
  Pen
)
}
\arguments{
\item{params}{A vector of parameters (gamma + b + sigma)}

\item{Y}{Survival time}

\item{delta}{Censoring indicator (1 = event)}

\item{X}{Design matrix}

\item{family}{"lognormal" or "loglogistic"}

\item{lambda}{A chosen nonnegative smoothing parameter}

\item{Pen}{A chosen penalty matrix}
}
\value{
A vector of negative penalized scores
}
\description{
Function to compute the gradient (score equations)
}
