% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil.R
\docType{data}
\name{pupil}
\alias{pupil}
\title{Pupil Light Response Dataset}
\format{
A data frame with 123 rows and 125 variables.
}
\source{
Internal dataset used in funAFT examples.
}
\usage{
data(pupil)
}
\description{
A dataset containing pupil light–response curves and cannabis user group information
for 123 subjects. Each row represents one subject.
}
\details{
The dataset includes:
\itemize{
\item \code{ptid}:
Patient ID.
\item \code{eye}:
The eye being tested (\code{Right}).
\item \code{time_since_use}:
Survival outcome. Time since most recent cannabis use (in minutes).
\item \code{is_user}:
Event indicator. \verb{1 = observed user}, \code{0 = censored}.
\item \code{pct_chg_1} to \code{pct_chg_119}:
Percent change in right-eye pupil diameter at 119 uniformly spaced time points
during a light-stimulus test. Each row represents the functional trajectory
\( X_i(s) \).
\item \code{age_in_years}:
Subject age (years).
\item \code{bmi}:
Body mass index.
}
}
\examples{
data(pupil)
str(pupil)

}
\keyword{datasets}
