% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty_matrix.R
\name{penalty_matrix}
\alias{penalty_matrix}
\title{Function to construct the penalty matrix}
\usage{
penalty_matrix(kp, nS, a)
}
\arguments{
\item{kp}{Number of basis coefficients}

\item{nS}{Density of the grid}

\item{a}{Parameter to ensure the penalty matrix is full rank (a < 0.01)}
}
\value{
A kp x kp penalty matrix
}
\description{
Function to construct the penalty matrix
}
